All the Jumps
Okay maybe not all the jumps but a bunch more condition jumps are available to us if I'd actually read the datasheet of our ALU chips. Implemented JG (jump if greater than X > Rb for Ra, Rc & Rd), JLE (jump if less than or equal), JGE (jump if greater than or equal), JL (jump if less than) and finally the JEQ (jump if equal). May have an issue with the other jumps which is commented on the sheet.