Fixed output hardware and software comunication
Didn't help that my wiring was incorrect and I had made a logic error in my circuit but ... all fixed now. Code was very easy to fix as was the wiring once the debug LEDs came out. All code changes commented. We have an unused section of the EEPROM that we could do something with, namely when Hexadecimal mode is 1 and two's complement is 1. Maybe another different mode is possible?