Fixed STR and LOD and wrote a test program
In attempting to write the first program, I realised our STR and LOD had no address parameter! Fixed them by adding all combinations of register as well as adding a jump if not equal to instruction as this was needed in the program. Would recommend trying to write your own program in our assembly language, it will show up ares of fault very quickly.